+++ pallets/amm/src/mock.rs
@@ -62,7 +62,9 @@ frame_support::construct_runtime!(
         Balances: pallet_balances::{Pallet, Call, Storage, Event<T>},
         Tokens: orml_tokens::{Pallet, Storage, Config<T>, Event<T>},
         Currencies: orml_currencies::{Pallet, Call, Event<T>},
-        AMM: pallet_amm::{Pallet, Call, Storage, Event<T>},
+        AMM: pallet_amm::<Instance1>::{Pallet, Call, Storage, Event<T>},
+        PermissionedAMM: pallet_amm::<Instance2>::{Pallet, Call, Storage, Event<T>},
+        DefaultAMM: pallet_amm::{Pallet, Call, Storage, Event<T>},
     }
 );
 
@@ -154,6 +156,28 @@ impl orml_currencies::Config for Test {
 
 parameter_types! {
     pub const AMMPalletId: PalletId = PalletId(*b"par/ammp");
+    pub const AllowPermissionlessPoolCreation: bool = true;
+}
+
+impl pallet_amm::Config<pallet_amm::Instance1> for Test {
+    type Event = Event;
+    type Currency = Currencies;
+    type PalletId = AMMPalletId;
+    type WeightInfo = ();
+    type AllowPermissionlessPoolCreation = AllowPermissionlessPoolCreation;
+}
+
+parameter_types! {
+    pub const PermissionedAMMPalletId: PalletId = PalletId(*b"par/ampe");
+    pub const ForbidPermissionlessPoolCreation: bool = false;
+}
+
+impl pallet_amm::Config<pallet_amm::Instance2> for Test {
+    type Event = Event;
+    type Currency = Currencies;
+    type PalletId = PermissionedAMMPalletId;
+    type WeightInfo = ();
+    type AllowPermissionlessPoolCreation = ForbidPermissionlessPoolCreation;
 }
 
 impl pallet_amm::Config for Test {
@@ -161,6 +185,7 @@ impl pallet_amm::Config for Test {
     type Currency = Currencies;
     type PalletId = AMMPalletId;
     type WeightInfo = ();
+    type AllowPermissionlessPoolCreation = AllowPermissionlessPoolCreation;
 }
 
 // Build genesis storage according to the mock runtime.